stannp/requirements.txt
requests
argparse
python-dotenv
